Update Group project phase III
I think I made a lot of progress. I changed the code to be more object-oriented. I made it so the ball bounces off the paddle (it wasn't doing that when I tested it before), breaks bricks, and ends if it touches the ground 3 times. I think as a last step we need to make the game restart when you press the restart button.